Use uintptr_t instead of uint32_t for addresses

+++ arch/x86/memory_dump.c
@@ -1,7 +1,7 @@
 #include <kernel/kernel.h>
 #include <kernel/backtrace.h>
 
-void memory_dump_impl(loglevel_t severity, const uint32_t* addr, size_t count)
+void memory_dump_impl(loglevel_t severity, const uintptr_t* addr, size_t count)
 {
     for (uint32_t i = 0; i < count; ++i)
     {
